Clean up aborted createWants stream
Otherwise, a createWants stream only gets cleaned up in the sink for the peer's
createWants stream. But for a local RPC connection, we do not call createWants
on the peer, so if the remote calls createWants, we have to clean it up after it
is aborted. Here we do that it in the appropriate place.